+++ crates/holochain/src/core/ribosome/host_fn/remote_signal.rs
@@ -51,7 +51,8 @@ pub fn remote_signal(
                 "remote_signal".into(),
             )
             .to_string(),
-        )).into()),
+        ))
+        .into()),
     }
 }
 
@@ -69,8 +70,8 @@ mod tests {
     fn zome(agents: Vec<AgentPubKey>, num_signals: Arc<AtomicUsize>) -> InlineZomeSet {
         let entry_def = EntryDef::default_with_id("entrydef");
 
-        SweetEasyInline::new(vec![entry_def.clone()], 0)
-            .callback("signal_others", move |api, ()| {
+        SweetInlineZomes::new(vec![entry_def.clone()], 0)
+            .function("signal_others", move |api, ()| {
                 let signal = ExternIO::encode("Hey").unwrap();
                 let signal = RemoteSignal {
                     agents: agents.clone(),
@@ -80,12 +81,12 @@ mod tests {
                 api.remote_signal(signal)?;
                 Ok(())
             })
-            .callback("recv_remote_signal", move |api, signal: ExternIO| {
+            .function("recv_remote_signal", move |api, signal: ExternIO| {
                 tracing::debug!("remote signal");
                 num_signals.fetch_add(1, Ordering::SeqCst);
                 api.emit_signal(AppSignal::new(signal)).map_err(Into::into)
             })
-            .callback("init", move |api, ()| {
+            .function("init", move |api, ()| {
                 let mut functions: GrantedFunctions = BTreeSet::new();
                 functions.insert((api.zome_info(()).unwrap().name, "recv_remote_signal".into()));
                 let cap_grant_entry = CapGrantEntry {
@@ -142,7 +143,7 @@ mod tests {
 
         let _: () = conductors[0]
             .call(
-                &cells[0].zome(SweetEasyInline::COORDINATOR),
+                &cells[0].zome(SweetInlineZomes::COORDINATOR),
                 "signal_others",
                 (),
             )